样式大改版，一些简单的优化

+++ src/main/java/com/homethy/controller/UserManagerController.java
@@ -2,10 +2,16 @@ package com.homethy.controller;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.homethy.UserHolder;
+import com.homethy.constant.Constant;
 import com.homethy.domain.DatabaseUserInfo;
+import com.homethy.domain.Menu;
+import com.homethy.domain.UserRole;
+import com.homethy.service.MenuService;
+import com.homethy.service.UserRoleService;
 import com.homethy.service.UserService;
-import com.homethy.util.DateUtil;
-
+import com.homethy.util.*;
+import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang.math.NumberUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -14,17 +20,25 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
-import java.util.Date;
-import java.util.List;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.*;
+import java.util.stream.Collectors;
 
 @Controller
-public class UserManagerController {
+public class UserManagerController extends BaseController{
 
   @Autowired
   private UserHolder userHolder;
 
   @Autowired
   private UserService userService;
+
+  @Autowired
+  private MenuService menuService;
+
+  @Autowired
+  private UserRoleService userRoleService;
 
   @RequestMapping(value = "update_password", method = RequestMethod.GET)
   public ModelAndView updatePassword() {
@@ -57,18 +71,29 @@ public class UserManagerController {
     return userService.updateUserPassword(password,newPassword,confirmPassword);
   }
 
-
-  @RequestMapping(value = "create_user", method = RequestMethod.GET)
-  public ModelAndView createUser() {
-    ModelAndView view = new ModelAndView("page_not_found");
-    if(userHolder.getUserInfo().getLevel()==9){
-      view = new ModelAndView("create_user");
+  @RequestMapping(value = "updateEmail", method = RequestMethod.POST)
+  @ResponseBody
+  public String updateEmail(@RequestParam(value = "email") String email){
+    DatabaseUserInfo userInfo = userHolder.getUserInfo();
+    if(email.equals(userInfo.getEmail())){
+        return ReturnJacksonUtil.resultFail("邮箱无修改");
     }
-    view.addObject("env",userHolder.getENV());
-    view.addObject("user",userHolder.getUserInfo());
-    view.addObject("databaseEnvList",userHolder.getDatabaseEnv());
-    return view;
+    userInfo.setEmail(email);
+    return userService.updateUserInfo(userInfo);
   }
+
+
+//  @RequestMapping(value = "create_user", method = RequestMethod.GET)
+//  public ModelAndView createUser() {
+//    ModelAndView view = new ModelAndView("page_not_found");
+//    if(userHolder.getUserInfo().getLevel()==9){
+//      view = new ModelAndView("create_user");
+//    }
+//    view.addObject("env",userHolder.getENV());
+//    view.addObject("user",userHolder.getUserInfo());
+//    view.addObject("databaseEnvList",userHolder.getDatabaseEnv());
+//    return view;
+//  }
 
   @ResponseBody
   @RequestMapping(value = "createUserSubmit", method = RequestMethod.POST)
@@ -76,30 +101,53 @@ public class UserManagerController {
       @RequestParam(value =  "account", required = true) String account,
       @RequestParam(value = "password", required = true) String password,
       @RequestParam(value = "confirmPassword", required = true) String  confirmPassword,
+      @RequestParam(value = "email", required = true) String email,
       @RequestParam(value = "level", required = false,defaultValue = "0") int level,
       @RequestParam(value = "onlineLevel", required = false, defaultValue = "0") int onlineLevel,
-    @RequestParam(value = "defaultSchema", required = false, defaultValue = "sitebuilt") String defaultSchema) throws JsonProcessingException {
-    return userService.createUserSubmit(account,password,confirmPassword,level,onlineLevel,defaultSchema);
+    @RequestParam(value = "defaultSchema", required = false, defaultValue = "sitebuilt") String defaultSchema,
+      @RequestParam(value = "roleGroup",defaultValue = "2") int roleGroup) throws JsonProcessingException {
+
+    if(!isDbAdminUser()){
+      return PERMISSION_DENIED;
+    }
+    if(!(password.equals(confirmPassword))){
+      return HomethyStringUtil.returnResultStr(Constant.ERROR,"2次输入的密码不匹配");
+    }
+    DatabaseUserInfo userInfo = new DatabaseUserInfo();
+    userInfo.setAccount(account);
+    userInfo.setPassword(MD5Support.hex(password, Constant.MD5KEY));
+    userInfo.setLevel(level);
+    userInfo.setOnlineLevel(onlineLevel);
+    userInfo.setDefaultSchema(defaultSchema);
+    userInfo.setStatus(1);
+    userInfo.setEmail(email);
+    return userService.createUserSubmit(userInfo,roleGroup);
   }
 
   @RequestMapping(value = "userManager", method = RequestMethod.GET)
   public ModelAndView userManager() {
     ModelAndView view = new ModelAndView("page_not_found");
-    if(userHolder.getUserInfo().getLevel()==9){
+    if(isDbAdminUser()){
       view = new ModelAndView("userManager");
     }
     view.addObject("env",userHolder.getENV());
     view.addObject("user",userHolder.getUserInfo());
     view.addObject("databaseEnvList",userHolder.getDatabaseEnv());
     view.addObject("schemaList",userHolder.getSchemaList());
+    view.addObject("roleList",userRoleService.getUserRoleList(100,1));
     return view;
   }
 
   @ResponseBody
   @RequestMapping(value = "getUserInfo", method = RequestMethod.GET)
   public String getUserInfo(@RequestParam(value = "limit") int limit,
-    @RequestParam(value = "offset") int offset,
-    @RequestParam(value = "key", required = false, defaultValue = "") String userName) {
+                            @RequestParam(value = "offset") int offset,
+                            @RequestParam(value = "key", required = false, defaultValue = "") String userName,
+                            HttpServletResponse response) throws IOException{
+    if(!isDbAdminUser()){
+      response.sendRedirect("/page_not_found");
+      return "";
+    }
     return userService.queryDatabaseUserInfo(limit, offset, userName);
   }
 
@@ -108,28 +156,45 @@ public class UserManagerController {
   public String updateUserInfo(@RequestParam(value = "id") int id,
     @RequestParam(value = "account") String account,
     @RequestParam(value = "level") int level,
-    @RequestParam(value = "online_level") int online_level,
-    @RequestParam(value = "status") int status,
-    @RequestParam(value = "default_schema") String default_schema) {
+    @RequestParam(value = "onlineLevel") int onlineLevel,
+    @RequestParam(value = Constant.RESULT_STATUS) int status,
+    @RequestParam(value = "defaultSchema") String defaultSchema,
+    @RequestParam(value = "email") String email,
+                               HttpServletResponse response) throws IOException{
+    if(!isDbAdminUser()){
+      response.sendRedirect("/page_not_found");
+      return "";
+    }
     DatabaseUserInfo databaseUserInfo = new DatabaseUserInfo();
     databaseUserInfo.setId(id);
     databaseUserInfo.setAccount(account);
     databaseUserInfo.setLevel(level);
-    databaseUserInfo.setOnlineLevel(online_level);
+    databaseUserInfo.setOnlineLevel(onlineLevel);
     databaseUserInfo.setStatus(status);
-    databaseUserInfo.setDefaultSchema(default_schema);
+    databaseUserInfo.setDefaultSchema(defaultSchema);
+    databaseUserInfo.setEmail(email);
     return userService.updateUserInfo(databaseUserInfo);
   }
 
   @ResponseBody
   @RequestMapping(value = "update/defaultSchema", method = RequestMethod.GET)
-  public String updateDefaultSchema(@RequestParam(value = "defaultSchema") String defaultSchema) {
+  public String updateDefaultSchema(@RequestParam(value = "defaultSchema") String defaultSchema,
+                                    HttpServletResponse response) throws IOException{
+    if(!isDbAdminUser()){
+      response.sendRedirect("/page_not_found");
+      return "";
+    }
     return userService.updateDefaultSchema(defaultSchema);
   }
 
   @ResponseBody
   @RequestMapping(value = "deleteUser", method = RequestMethod.GET)
-  public String deleteUser(@RequestParam(value = "idList") String idList) {
+  public String deleteUser(@RequestParam(value = "idList") String idList,
+                           HttpServletResponse response) throws IOException{
+    if(!isDbAdminUser()){
+      response.sendRedirect("/page_not_found");
+      return "";
+    }
     return userService.deleteUser(idList);
   }
 
@@ -137,6 +202,77 @@ public class UserManagerController {
   @RequestMapping(value = "getSqlDetail", method = RequestMethod.GET)
   public String getSqlDetail(@RequestParam(value = "sqlId") int sqlId) {
     return userService.getSqlDetail(sqlId);
+  }
+
+  @ResponseBody
+  @RequestMapping(value = "role/getUserRoleList", method = RequestMethod.GET)
+  public String getUserRoleList(HttpServletResponse response,@RequestParam(value = "limit") int limit,
+                                @RequestParam(value = "page")int page) throws IOException{
+    if(!isDbAdminUser()){
+      response.sendRedirect("/page_not_found");
+      return "";
+    }
+
+    List<UserRole> list = userRoleService.getUserRoleList(limit,page);
+    Map result = new HashMap();
+    result.put(Constant.RESULT_DATA,list);
+    result.put(Constant.RESULT_CODE,0);
+    result.put("msg","");
+    result.put("count",list.size());
+    return JacksonUtils.toJson(result);
+  }
+
+  @RequestMapping(value = "role/edit", method = RequestMethod.GET)
+  public ModelAndView roleEdit(HttpServletResponse response,@RequestParam(value = "id",defaultValue = "0") int id) {
+
+    if(!isDbAdminUser()){
+      return new ModelAndView("page_not_found");
+    }
+
+    ModelAndView view = new ModelAndView("/winUi/role/edit");
+
+    UserRole userRole = new UserRole();
+    if(id != 0){
+      userRole = userRoleService.getUserRoleById(id);
+    }
+
+    List<Menu> allMenus = menuService.getAllMenuByRole(id);
+
+    view.addObject("sysMenus",allMenus);
+    view.addObject("userRole",userRole);
+    view.addObject("dbUser",userHolder.getUserInfo());
+
+    return view;
+  }
+
+  @RequestMapping(value = "role/update", method = RequestMethod.POST)
+  @ResponseBody
+  public String roleUpdate(HttpServletResponse response,@RequestParam(value = "roleName") String roleName,
+                           @RequestParam(value = "description") String description,
+                           @RequestParam(value = "menus") String menus,@RequestParam(value = "id" ,defaultValue = "0")int id) throws IOException{
+
+    if(!isDbAdminUser()){
+      response.sendRedirect("/page_not_found");
+      return "";
+    }
+    UserRole userRole = new UserRole(id,roleName,description);
+    //插入，得到id
+    if(id == 0){
+      id = userRoleService.insertUserRole(userRole);
+      userRole.setId(id);
+    }else{
+      userRoleService.updateUserRole(userRole);
+    }
+    List<Integer> list = null;
+    if(StringUtils.isNotBlank(menus)){
+      list = Arrays.stream(menus.substring(0, menus.length() - 1).split(",")).map(roleId -> {
+        return NumberUtils.toInt(roleId);
+      }).collect(Collectors.toList());
+    }
+
+    int n = menuService.updateRoleMenus(id,list);
+
+    return String.valueOf(n);
   }
 
 }
